add UUIDs and remove orphan file tags
