+++ designsafe/apps/projects/models/agave/base.py
@@ -14,6 +14,16 @@ from designsafe.apps.data.models.elasticsearch import IndexedPublication, Indexe
 from designsafe.libs.elasticsearch.exceptions import DocumentNotFound
 logger = logging.getLogger(__name__)
 
+class FileObjModel(MetadataModel):
+    _is_nested = True
+    path = fields.CharField('File path', max_length=1024, default='')
+    name = fields.CharField('File name', max_length=1024, default='')
+    system = fields.CharField('System', max_length=512, default='')
+    type = fields.CharField('File type', max_length=512, default='')
+    last_modified = fields.CharField('Last Modified', max_length=512, default='')
+    uuid = fields.CharField('UUID', max_length=512, default='')
+    length = fields.IntField('File Size', default=0)
+
 
 class RelatedEntity(MetadataModel):
     """Model for entities related to projects."""
@@ -25,6 +35,17 @@ class RelatedEntity(MetadataModel):
         for attrname, field in six.iteritems(self._meta._related_fields):
             body_dict['_relatedFields'].append(attrname)
         return body_dict
+    
+    def save(self, ac):
+        """Remove file tags if the file has been un-associated from the entity."""
+        file_objs = getattr(self, 'file_objs', [])
+        file_tags = getattr(self, 'file_tags', [])
+
+        if file_objs and file_tags:
+            filtered_tags = filter(lambda tag: bool([file for file in file_objs if tag["path"].startswith(file["path"])]), file_tags)
+            self.file_tags = list(filtered_tags)
+        
+        return super().save(ac)
 
     def to_datacite_json(self):
         """Serialize object to datacite JSON.
