+++ designsafe/apps/projects/models/agave/rapid.py
@@ -3,7 +3,7 @@ import logging
 import json
 from designsafe.apps.data.models.agave.base import Model as MetadataModel
 from designsafe.apps.data.models.agave import fields
-from designsafe.apps.projects.models.agave.base import RelatedEntity, Project
+from designsafe.apps.projects.models.agave.base import RelatedEntity, Project, FileObjModel
 
 logger = logging.getLogger(__name__)
 
@@ -57,15 +57,6 @@ class DataTag(MetadataModel):
     _is_nested = True
     file = fields.RelatedObjectField(FileModel, default=[])
     desc = fields.CharField('Description', max_length=512, default='')
-
-class FileObjModel(MetadataModel):
-    _is_nested = True
-    path = fields.CharField('File path', max_length=1024, default='')
-    name = fields.CharField('File name', max_length=1024, default='')
-    system = fields.CharField('System', max_length=512, default='')
-    type = fields.CharField('File type', max_length=512, default='')
-    length = fields.IntField('File Size', default=0)
-    last_modified = fields.CharField('Last Modified', max_length=512, default='')
 
 
 class Mission(RelatedEntity):
