+++ designsafe/static/scripts/projects/components/file-categories/file-categories.component.js
@@ -102,6 +102,7 @@ class FileCategoriesCtrl {
             }
             this.rmUuid(entity.associationIds, this.file.uuid);
             this.rmUuid(entity.value.files, this.file.uuid);
+            entity.value.fileObjs = (entity.value.fileObjs ?? []).filter(f => f.path !== this.file.path)
             this.updateEntity(entity, this.file);
         }
     }
